+++ src/test/java/org/scify/jedai/blockprocessing/comparisoncleaning/TestAllMethods.java
@@ -53,7 +53,7 @@ public class TestAllMethods {
         System.out.println("Existing Duplicates\t:\t" + duplicatePropagation.getDuplicates().size());
         
         for (BlockBuildingMethod blbuMethod : BlockBuildingMethod.values()) {
-            double time1 = System.currentTimeMillis();
+            float time1 = System.currentTimeMillis();
             
             StringBuilder workflowConf = new StringBuilder();
             StringBuilder workflowName = new StringBuilder();
@@ -81,7 +81,7 @@ public class TestAllMethods {
                 workflowName.append("->").append(comparisonCleaningMethod.getMethodName());
             }
             
-            double time2 = System.currentTimeMillis();
+            float time2 = System.currentTimeMillis();
             
             BlocksPerformance blStats = new BlocksPerformance(blocks, duplicatePropagation);
             blStats.setStatistics();
