+++ src/test/java/org/scify/jedai/configuration/ConfigureRandomlyEndToEndWorkflowCCER.java
@@ -100,9 +100,9 @@ public class ConfigureRandomlyEndToEndWorkflowCCER {
             matchingWorkflowName.append("->").append(ec.getMethodName());
 
             int bestIteration = 0;
-            double bestFMeasure = 0;
+            float bestFMeasure = 0;
             for (int j = 0; j < NO_OF_TRIALS; j++) {
-                double time1 = System.currentTimeMillis();
+                float time1 = System.currentTimeMillis();
 
                 bb.setNextRandomConfiguration();
                 final List<AbstractBlock> blocks = bb.getBlocks(profiles1, profiles2);
@@ -131,7 +131,7 @@ public class ConfigureRandomlyEndToEndWorkflowCCER {
                 ec.setNextRandomConfiguration();
                 final EquivalenceCluster[] clusters = ec.getDuplicates(sims);
 
-                double time2 = System.currentTimeMillis();
+                float time2 = System.currentTimeMillis();
 
                 final StringBuilder matchingWorkflowConf = new StringBuilder();
                 matchingWorkflowConf.append(bb.getMethodConfiguration());
@@ -145,7 +145,7 @@ public class ConfigureRandomlyEndToEndWorkflowCCER {
                 clp.setStatistics();
                 clp.printStatistics(time2 - time1, matchingWorkflowName.toString(), matchingWorkflowConf.toString());
 
-                double fMeasure = clp.getFMeasure();
+                float fMeasure = clp.getFMeasure();
                 if (bestFMeasure < fMeasure) {
                     bestIteration = j;
                     bestFMeasure = fMeasure;
@@ -155,7 +155,7 @@ public class ConfigureRandomlyEndToEndWorkflowCCER {
             System.out.println("\nBest Iteration\t:\t" + bestIteration);
             System.out.println("Best FMeasure\t:\t" + bestFMeasure);
 
-            double time1 = System.currentTimeMillis();
+            float time1 = System.currentTimeMillis();
 
             bb.setNumberedRandomConfiguration(bestIteration);
             final List<AbstractBlock> blocks = bb.getBlocks(profiles1, profiles2);
@@ -175,7 +175,7 @@ public class ConfigureRandomlyEndToEndWorkflowCCER {
             ec.setNumberedRandomConfiguration(bestIteration);
             final EquivalenceCluster[] clusters = ec.getDuplicates(sims);
 
-            double time2 = System.currentTimeMillis();
+            float time2 = System.currentTimeMillis();
 
             final StringBuilder matchingWorkflowConf = new StringBuilder();
             matchingWorkflowConf.append(bb.getMethodConfiguration());
