+++ src/test/java/org/scify/jedai/configuration/HolisticRandomConfigurationCCER.java
@@ -97,9 +97,9 @@ public class HolisticRandomConfigurationCCER {
             matchingWorkflowName.append("->").append(ec.getMethodName());
 
             int bestIteration = 0;
-            double bestFMeasure = 0;
+            float bestFMeasure = 0;
             for (int j = 0; j < NO_OF_TRIALS; j++) {
-                double time1 = System.currentTimeMillis();
+                float time1 = System.currentTimeMillis();
 
                 bb.setNextRandomConfiguration();
                 final List<AbstractBlock> blocks = bb.getBlocks(profiles1, profiles2);
@@ -128,7 +128,7 @@ public class HolisticRandomConfigurationCCER {
                 ec.setNextRandomConfiguration();
                 final EquivalenceCluster[] clusters = ec.getDuplicates(sims);
 
-                double time2 = System.currentTimeMillis();
+                float time2 = System.currentTimeMillis();
 
                 final StringBuilder matchingWorkflowConf = new StringBuilder();
                 matchingWorkflowConf.append(bb.getMethodConfiguration());
@@ -142,7 +142,7 @@ public class HolisticRandomConfigurationCCER {
                 clp.setStatistics();
                 clp.printStatistics(time2 - time1, matchingWorkflowName.toString(), matchingWorkflowConf.toString());
 
-                double fMeasure = clp.getFMeasure();
+                float fMeasure = clp.getFMeasure();
                 if (bestFMeasure < fMeasure) {
                     bestIteration = j;
                     bestFMeasure = fMeasure;
@@ -152,7 +152,7 @@ public class HolisticRandomConfigurationCCER {
             System.out.println("\nBest Iteration\t:\t" + bestIteration);
             System.out.println("Best FMeasure\t:\t" + bestFMeasure);
 
-            double time1 = System.currentTimeMillis();
+            float time1 = System.currentTimeMillis();
 
             bb.setNumberedRandomConfiguration(bestIteration);
             final List<AbstractBlock> blocks = bb.getBlocks(profiles1, profiles2);
@@ -172,7 +172,7 @@ public class HolisticRandomConfigurationCCER {
             ec.setNumberedRandomConfiguration(bestIteration);
             final EquivalenceCluster[] clusters = ec.getDuplicates(sims);
 
-            double time2 = System.currentTimeMillis();
+            float time2 = System.currentTimeMillis();
 
             final StringBuilder matchingWorkflowConf = new StringBuilder();
             matchingWorkflowConf.append(bb.getMethodConfiguration());
